refactor(discord): split components into base + per-game provider package (#268)
components.py -> components/{base,dota,deadlock,default,registry}.py + __init__
re-exports. DotaComponents/DeadlockComponents/DefaultComponents own
build_signup_modal; SignupButton dispatches via get_component_provider().
Codec-based custom_ids (byte-identical). Bug-2 defer-first on the ORM-bearing
Dota callbacks (position_confirm/rank_star_select/battle_cup_select):
response.defer() before the ORM call, edit_original_response after. Registry
falls back to DefaultComponents with a loud provider_fallback log.
Full discordbot.tests suite: 152 pass.
